+++ PoemEngineUi/PopupPoem.qml
@@ -15,12 +15,37 @@ Popup {
     property string yan: ""
     property string title: ""
     property string ticai: ""
+    property string yun: ""
     property string ju: ""
     property string juind: ""
     property string pz: ""
     property var poem: null
-    property var yuns: null
+    property var psy: interf.psy
     property string zi_disp: ""
+    property int lineSelected : -1
+    property int ziSelected : -1
+    property var zi_yuns: {
+        let ret = []
+        for(let yunname of poem["韻目"][root.zi_disp]) {
+            ret.push(root.psy[yunname])
+        }
+        return ret
+    }
+
+    function get_zi_pz(zi) {
+        let pz_ret = "？"
+        for(let yunname of poem["韻目"][zi]) {
+            let yun = root.psy[yunname]
+            // console.error("韵", yunname, JSON.stringify(yun), poem["韻目"][zi])
+            if(pz_ret === "？")
+                pz_ret = get_pz(yun);
+            else if(pz_ret !== "通" && pz_ret !== get_pz(yun)){
+                return "通"
+            }
+        }
+        return pz_ret
+    }
+
     property var ju_list: {
         const lines = []
         const regex = /[^，。？！；]+[，。？！；]/g
@@ -112,26 +137,49 @@ Popup {
                     model: root.ju_list
 
                     delegate: Flow {
+                        id: lineItem
+                        property int lineIndex: index
                         spacing: 0
-                        Layout.maximumWidth: root.availableWidth
+                        Layout.maximumWidth: root.availableWidth - 10
                         Layout.alignment: Qt.AlignHCenter
+                        Layout.leftMargin: 10
+                        Layout.rightMargin: 10
 
                         Repeater {
                             model: modelData
 
                             delegate: Item {
+                                property int i: lineIndex
+                                property int j: index
                                 implicitWidth: textItem.implicitWidth
                                 implicitHeight: textItem.implicitHeight
                                 Layout.alignment: Qt.AlignHCenter
 
-                                Text {
+                                // Text {
+                                //     id: textItem
+                                //     anchors.fill: parent
+                                //     font.pixelSize: 16
+                                //     color: Material.foreground
+                                //     horizontalAlignment: Text.AlignHCenter
+                                //     text: modelData[0]
+                                //     font.bold: modelData[1]
+                                // }
+
+                                PzTextItem {
                                     id: textItem
                                     anchors.fill: parent
-                                    font.pixelSize: 16
-                                    color: Material.foreground
-                                    horizontalAlignment: Text.AlignHCenter
-                                    text: modelData[0]
-                                    font.bold: modelData[1]
+                                    fontSize: 16
+                                    fontBold: modelData[1]
+                                    ju: modelData[0]
+                                    pz: get_zi_pz(modelData[0])
+                                }
+
+                                Rectangle {
+                                    z: -1
+                                    color: "#DDDDDD"
+                                    anchors.fill: parent
+                                    radius: 10
+                                    visible: root.lineSelected === i && root.ziSelected === j
                                 }
 
                                 MouseArea {
@@ -156,7 +204,11 @@ Popup {
                                         else
                                             root.zi_disp = ""
 
-                                        interf.searchYunsByZi(modelData[0])
+                                        // interf.searchYunsByZi(modelData[0])
+
+                                        root.lineSelected = i
+                                        root.ziSelected = j
+                                        console.error("Selected:", i, j)
                                     }
                                 }
                             }
@@ -167,12 +219,19 @@ Popup {
             }
         }
 
+        Rectangle {
+            Layout.fillWidth: true
+            Layout.preferredHeight: 1
+            color: Material.theme === Material.Light ? "#ccc" : "#444"
+            visible: poem["注釋"] !== ""
+        }
+
         RowLayout {
             Label {
-                text: qsTr("体裁：") + root.yan === -1 ? root.yan : "杂" + qsTr("言 ")
+                Component.onCompleted: {console.error("体裁", implicitHeight, implicitWidth) }
+                text: /*qsTr("体裁：") +*/ (root.yan !== "-1" ? root.yan : qsTr("杂")) + qsTr("言 ")
                       + root.jushu + qsTr("句 ")
                       + root.ticai
-                rightPadding: 10
                 font.pixelSize: 14
                 horizontalAlignment: Text.AlignLeft
                 Layout.alignment: Qt.AlignRight
@@ -180,9 +239,31 @@ Popup {
                 wrapMode: Text.Wrap
             }
 
+            PzTextItem {
+                Component.onCompleted: {console.error("韵脚", implicitHeight, implicitWidth) }
+                property var yunbu: root.psy[root.yun]
+                ju: {
+                    if (root.yun === "")
+                        return qsTr("无")
+
+                    return yunbu["聲調"] + yunbu["序號"] + yunbu["韻名"]
+                }
+                pz: {
+                    let pz = get_pz(yunbu)
+                    let len = ju.length
+                    if (len < 3) return pz.repeat(len)
+                    return pz.repeat(2) + "平".repeat(len - 3) + pz
+                }
+
+                fontBold: false
+                fontSize: 14
+                textHorizontalAlignment: Text.AlignHCenter
+                verticalMargin: 0
+            }
+
             Label {
-                text: qsTr("　出律度：") + parseFloat(root.poem["出律度"]).toFixed(2)
-                rightPadding: 10
+                Component.onCompleted: {console.error("出律", implicitHeight, implicitWidth) }
+                text: qsTr("出律度") + parseFloat(root.poem["出律度"]).toFixed(2)
                 font.pixelSize: 14
                 horizontalAlignment: Text.AlignRight
                 Layout.alignment: Qt.AlignRight
@@ -248,33 +329,24 @@ Popup {
         }
 
         RowLayout {
-            Layout.preferredHeight: 60
+            Layout.preferredHeight: 30
             Layout.minimumHeight: yunsFlow.implicitHeight
             visible: root.zi_disp !== ""
             // Layout.preferredHeight: 60
             PzTextItem {
+                // DebugOutline {}
                 id: item_zi
-                Layout.fillHeight: true
+                // Layout.fillHeight: true
                 Layout.alignment: Qt.AlignHCenter
                 Layout.minimumWidth: implicitWidth
                 Layout.maximumWidth: 40
                 Layout.preferredWidth: implicitWidth
+                Layout.preferredHeight: 30
                 Layout.leftMargin: 10
                 Layout.rightMargin: 10
                 fontSize: 24
                 ju: root.zi_disp
-                pz: {
-                    let pz_ret = "？"
-                    for(let yun of root.yuns) {
-                        console.error("韵", yun, root.yuns)
-                        if(pz_ret === "？")
-                            pz_ret = get_pz(yun);
-                        else if(pz_ret !== "通" && pz_ret !== get_pz(yun)){
-                            return "通"
-                        }
-                    }
-                    return pz_ret;
-                }
+                pz: get_zi_pz(root.zi_disp)
                 // {
                 //     // console.error("QML DEBUG:", "item_zi:", root.zi_disp == "")
                 //     return root.zi_disp !== ""
@@ -282,20 +354,21 @@ Popup {
             }
 
             Flow {
+                // DebugOutline {}
                 id: yunsFlow
                 Layout.fillWidth: true
+                Layout.preferredHeight: implicitHeight
                 // Layout.fillHeight: true
                 Layout.alignment: Qt.AlignHCenter | Qt.AlignVCenter
                 spacing: 10
 
                 Repeater {
-                    model: root.yuns
-
+                    model: zi_yuns
                     delegate: Item {
                         implicitWidth: yunText.implicitWidth + 10
                         implicitHeight: yunText.implicitHeight + 5
-                        Layout.topMargin: 2
-                        Layout.bottomMargin: 2
+                        // Layout.topMargin: 2
+                        // Layout.bottomMargin: 2
                         Layout.alignment: Qt.AlignLeft | Qt.AlignVCenter
                         Text {
                             id: yunText