+++ PoemEngineUi/PzTextItem.qml
@@ -11,6 +11,10 @@ Item {
     property string pz: "平仄平平仄？？通通仄平"
     property bool dispPz: true
     property int fontSize: 20
+    property bool fontBold: true
+    property int textHorizontalAlignment: Text.AlignHCenter
+    property int textVerticalAlignment: Text.AlignVCenter
+    property int verticalMargin: -2
 
     property var colors: Material.theme === Material.Light ? {
                             "平": "#3B3838",
@@ -27,8 +31,10 @@ Item {
     Text {
         id: lower
         textFormat: Text.RichText
+        horizontalAlignment: textHorizontalAlignment
+        // verticalAlignment: textVerticalAlignment
         font.pixelSize: root.fontSize
-        font.bold: true
+        font.bold: fontBold
         text: {
             if (!root.dispPz)
             {
@@ -46,14 +52,16 @@ Item {
         }
         anchors.left: parent.left
         anchors.top: parent.top
-        anchors.topMargin: (root.height-implicitHeight) / 2 - 2
+        anchors.topMargin: (root.height-implicitHeight) / 2 + verticalMargin
     }
 
     // 上半部分（红）
     Text {
         id: upper
         font.pixelSize: root.fontSize
-        font.bold: true
+        horizontalAlignment: textHorizontalAlignment
+        // verticalAlignment: textVerticalAlignment
+        font.bold: fontBold
         color: "#E57A31"
         clip: true
         height: implicitHeight/2
@@ -70,6 +78,6 @@ Item {
             return str
         }
         anchors.top: parent.top
-        anchors.topMargin: (root.height-implicitHeight) / 2 - 2
+        anchors.topMargin: (root.height-implicitHeight) / 2 + verticalMargin
     }
 }